+++ src/features/oauth/lib/useLoginOauth.ts
@@ -1,11 +1,8 @@
 'use client'
 
 import { useNavigater } from '@/shared/lib'
-import { useSearchParams } from 'next/navigation'
-import { useEffect } from 'react'
 
 const UseLoginOauth = () => {
-  const searchParams = useSearchParams()
   const { navigateOnboarding, navigateToKakaoAuth } = useNavigater()
 
   const loginController = () => {
@@ -17,15 +14,6 @@ const UseLoginOauth = () => {
       navigateToKakaoAuth()
     }
   }
-
-  useEffect(() => {
-    const accessToken = searchParams.get('token')
-    if (accessToken) {
-      localStorage.setItem('access_token', accessToken)
-      navigateOnboarding()
-    }
-  }, [])
-
   return { loginController }
 }
